liuhaoze/whisper-finetuning/whisper-finetuning/requirements.txt
torch
numpy
git+https://github.com/openai/whisper.git
tqdm
transformers
bitsandbytes
faiss-cpu